refactor: use product card everywhere
